solution/day06/src/elfWorkshop.ts
export class ElfWorkshop {
    taskList: string[] = [];

    // Adding tasks to the workshop is crucial for our workflow.
    // To enhance robustness, might we consider also checking 
    // for strings that are only whitespace?
    // This small check can ensure all tasks added hold meaningful content.
    addTask(task: string) {
        // It's great to see validation for non-empty tasks.
        // Perhaps, extending this to ignore tasks that only contain spaces (e.g., using task.trim() !== "")
        // could help maintain the task list's meaningfulness.
        if (task.trim() !== "") {
            this.taskList.push(task);
        }
    }

    // The approach to complete tasks is straightforward and effective.
    // I'm curious if returning a specific message or throwing an error when there are no tasks
    // left could provide clearer feedback to the elves or the system managing tasks.
    completeTask() {
        // I appreciate the simplicity here. For enhancing clarity for empty task lists,
        // might we consider an alternative return value or a specific message indicating
        // no tasks are available? It could enrich the method's communicative value.
        if (this.taskList.length > 0) {
            return this.taskList.shift();
        }
        return null;
    }
}